tegra: mmc: Honor max-frequency in device-tree
The driver currently hard-codes the max freqency for the sdhci controllers.
If the controller is unable to operate at the max frequency, the mmc card will not be available on the first scan.
Subsequent scans will eventually find a working combination.

Fix this by allowing the driver to check for the max-frequency handle and default to the original value if it doesn't exist.

Signed-off-by: Peter Geis <pgwipeout@gmail.com>